# backend/openapi.yaml
openapi: 3.1.0
info:
  title: AIBOM Model Catalog API
  version: 1.0.0
  description: |
    MySQL 기반 AI 모델 카탈로그 및 모델 계층(lineage) 데이터를 제공하는 FastAPI 백엔드 명세입니다.

    ## 구성
    - **system**: 서비스/DB 상태 확인
    - **families**: 모델 패밀리 인덱스와 선택 패밀리의 계층(hierarchy) 트리
    - **models**: 개별 모델 상세 메타데이터 (`model_info` + `model_hierarchy`)
    - **admin**: 모델 CRUD 및 OpenRouter 기반 AI 리서치 초안 생성

    ## 공통 규칙
    - 모든 JSON 필드는 **camelCase** 를 사용합니다. (families hierarchy 응답은 예외적으로 snake_case)
    - `modelId` 는 `namespace/model-name` 형식이며 슬래시를 포함하므로 경로 파라미터는 path 전체를 매칭합니다.
      예: `GET /api/v1/models/meta-llama/Llama-3.1-8B-Instruct`
    - DB 연결 실패 시 DB를 사용하는 모든 엔드포인트는 `503` 을 반환합니다.
    - 요청 본문 검증 실패(Pydantic) 시 FastAPI 표준 `422` 응답을 반환합니다.
    - 인증/인가는 현재 구현되어 있지 않습니다. admin 엔드포인트도 공개 상태입니다.
    - CORS: `CORS_ORIGINS` 환경 변수 (기본 `http://localhost:5173,http://127.0.0.1:5173`), 허용 메서드 GET/POST/PUT/DELETE

    ## 프론트엔드 로딩 순서
    1. `GET /api/v1/families` 로 경량 인덱스를 먼저 로드
    2. 선택된 패밀리에 대해 `GET /api/v1/families/{familyKey}/hierarchy` 지연 로드
    3. 노드 클릭 시 `GET /api/v1/models/{modelId}` 로 상세 로드
  contact:
    name: AIBOM PoC
servers:
  - url: http://127.0.0.1:8000
    description: Local development (uvicorn app.main:app --reload)
...
paths:
...
  /api/v1/families:
    get:
      tags: [families]
      summary: 패밀리 인덱스 조회
      description: UI 부트스트랩에 필요한 경량 패밀리 목록과 집계 요약을 반환합니다. 패밀리 출시일 오름차순, 동일 시 `familyKey` 순으로 정렬됩니다.
      operationId: listFamilies
      responses:
        '200':
          description: 패밀리 인덱스
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyIndex'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /api/v1/families/{familyKey}/hierarchy:
    get:
      tags: [families]
      summary: 패밀리 계층 트리 조회
      description: |
        선택한 패밀리의 루트 노드와 파생 모델 목록을 반환합니다.
        루트는 `sibling_order` 순, 파생 모델은 `(parent_model_id, sibling_order)` 순으로 정렬됩니다.
        `details_json.family_config` 가 있으면 그 값을 기반으로, 없으면 기본 문서 노드를 생성합니다.
        이 응답의 필드는 **snake_case** 입니다.
      operationId: getFamilyHierarchy
      parameters:
        - $ref: '#/components/parameters/FamilyKey'
      responses:
        '200':
          description: 패밀리 계층
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FamilyHierarchy'
        '404':
          description: 패밀리를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              example: { detail: Model family not found. }
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /api/v1/models:
    get:
      tags: [models]
      summary: 전체 모델 목록 조회
      description: 모든 모델의 상세 메타데이터를 반환합니다. 패밀리 출시일, `familyKey`, `modelId` 순으로 정렬됩니다.
      operationId: listModels
      responses:
        '200':
          description: 모델 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelDetail'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /api/v1/models/{modelId}:
    get:
      tags: [models]
      summary: 모델 상세 조회
      description: |
        `modelId` 는 `namespace/model-name` 형식이며 슬래시를 포함합니다 (path 파라미터 전체 매칭).
        카탈로그 레코드와 함께 설계안 8개 영역의 AIBOM을 `aibom` 에 담아 반환합니다 (저장된 AIBOM이 없으면 `null`).
      operationId: getModel
      parameters:
        - $ref: '#/components/parameters/ModelId'
      responses:
        '200':
          description: 모델 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelDetail'
        '404':
          description: 모델을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              example: { detail: Model not found. }
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /api/v1/admin/models:
    get:
      tags: [admin]
      summary: '[관리자] 모델 목록 조회'
      description: 관리 화면용 평탄화된(flat) 모델 레코드 목록입니다. `familyKey`, `modelId` 순으로 정렬됩니다. BOM/체크리스트 필드와 계층 정보를 모두 포함합니다.
      operationId: adminListModels
      responses:
        '200':
          description: 관리자용 모델 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminModel'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
    post:
      tags: [admin]
      summary: '[관리자] 모델 생성'
      description: |
        `model_info` 와 `model_hierarchy` 레코드를 함께 생성합니다.
        AIBOM 영역도 함께 초기화합니다: `provenance`(subject, origin=familyName, provider=supplier, parent, relation),
        부모가 있으면 `transformation`(input = 부모, role base), `model`(intended_use = primaryPurpose),
        `familyLicenseName` 이 있으면 `license_policy`.
        - `parentModelId` 가 있으면 부모가 존재하고 동일 패밀리여야 하며, `hierarchyDepth` 는 부모 depth + 1 로 계산됩니다.
        - 루트 모델(`parentModelId: null`)은 `relationshipType` 을 가질 수 없고, 파생 모델은 반드시 가져야 합니다.
      operationId: adminCreateModel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminModelWrite'
      responses:
        '201':
          description: 생성됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminModel'
        '409':
          description: 동일 `modelId` 가 이미 존재하거나 DB 무결성 제약 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                duplicate:
                  value: { detail: A model with this ID already exists. }
                integrity:
                  value: { detail: The model could not be saved because its data conflicts with another record. }
        '422':
          description: 검증 실패 (Pydantic 또는 계층 규칙)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDetail'
                  - $ref: '#/components/schemas/HTTPValidationError'
              examples:
                parentMissing:
                  value: { detail: The selected parent model does not exist. }
                familyMismatch:
                  value: { detail: Parent and child must belong to the same family. }
                noHierarchy:
                  value: { detail: The selected parent has no hierarchy record. }
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /api/v1/admin/models/{modelId}:
    put:
      tags: [admin]
      summary: '[관리자] 모델 수정'
      description: |
        모델 메타데이터와 계층 위치를 전체 교체(full replace)합니다.
        - 본문의 `modelId` 는 경로의 `modelId` 와 동일해야 합니다 (ID 변경 불가).
        - 자식이 있는 모델은 다른 패밀리로 이동할 수 없습니다.
        - 자기 자신 또는 자신의 후손을 부모로 지정하면 순환으로 거부됩니다.
        - 부모가 바뀌면 모든 후손의 `hierarchyDepth` 가 재계산됩니다.
        - 부모 또는 `relationshipType` 이 바뀐 경우에만 AIBOM의 `provenance.parent` / `relation` 과
          `transformation.input` 의 base 항목이 새 부모로 교체됩니다 (base가 아닌 input 은 유지).
      operationId: adminUpdateModel
      parameters:
        - $ref: '#/components/parameters/ModelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminModelWrite'
      responses:
        '200':
          description: 수정됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminModel'
        '404':
          description: 모델을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              example: { detail: 'Model not found: org/model' }
        '409':
          description: DB 무결성 제약 충돌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              example: { detail: The model could not be saved because its data conflicts with another record. }
        '422':
          description: 검증 실패 (Pydantic 또는 계층 규칙)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDetail'
                  - $ref: '#/components/schemas/HTTPValidationError'
              examples:
                idChanged:
                  value: { detail: Model ID cannot be changed after creation. }
                moveWithChildren:
                  value: { detail: A model with children cannot be moved to another family. }
                cycle:
                  value: { detail: The selected parent would create a hierarchy cycle. }
                parentMissing:
                  value: { detail: The selected parent model does not exist. }
                familyMismatch:
                  value: { detail: Parent and child must belong to the same family. }
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
    delete:
      tags: [admin]
      summary: '[관리자] 모델 삭제'
      description: |
        자식이 없는 모델만 삭제할 수 있습니다. `model_hierarchy` 행을 먼저 삭제한 뒤 `model_info` 행을 삭제합니다.
        이 모델의 AIBOM 행(`model`, `provenance`, `transformation`, `evaluation`, `safety_ethics`, `license_policy`)도 함께 삭제합니다.
        다른 모델의 `transformation.input` 으로 쓰이는 모델은 409로 거부됩니다.
      operationId: adminDeleteModel
      parameters:
        - $ref: '#/components/parameters/ModelId'
      responses:
        '204':
          description: 삭제됨 (본문 없음)
        '404':
          description: 모델을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              example: { detail: 'Model not found: org/model' }
        '409':
          description: 자식 모델이 존재함
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
              examples:
                children:
                  value: { detail: "Move or delete this model's children before deleting it." }
                aibomInput:
                  value: { detail: 'This model is an input of other transformations (org/child). Update their AIBOM first.' }
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /api/v1/admin/models/{modelId}/aibom:
    get:
      tags: [admin]
      summary: '[관리자] 모델 AIBOM 조회'
      description: 설계안 8개 영역의 AIBOM. 저장된 행이 없으면 빈 객체를 반환합니다.
      operationId: adminGetModelAibom
      parameters:
        - $ref: '#/components/parameters/ModelId'
      responses:
        '200':
          description: 모델 AIBOM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelAibom'
        '404':
          description: 모델을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
    put:
      tags: [admin]
      summary: '[관리자] 모델 AIBOM 저장'
      description: |
        이 모델의 AIBOM 행을 전체 교체합니다. `provenance.parent` 는 `transformation.input` 중 role이
        base / merge_source 인 항목으로 자동 계산됩니다. `dataset` / `reference` 항목은 공유 행으로 upsert 되며,
        `provenance.evidence` 에는 reference id 또는 URI(없으면 새로 등록)를 넣을 수 있습니다.
        `dataset.role` 은 카탈로그 전체의 사용처로 다시 계산됩니다.
      operationId: adminSaveModelAibom
      parameters:
        - $ref: '#/components/parameters/ModelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AibomWrite'
      responses:
        '200':
          description: 저장된 AIBOM
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelAibom'
        '404':
          description: 모델을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: 검증 실패 (없는 dataset / reference 참조, 자기 자신을 input으로 지정, 중복 license 등)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/ErrorDetail'
                  - $ref: '#/components/schemas/HTTPValidationError'
              example: { detail: 'Unknown dataset(s): org/data. Add them under `dataset` first.' }
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /api/v1/admin/aibom/options:
    get:
      tags: [admin]
      summary: '[관리자] AIBOM에서 참조 가능한 dataset / reference 목록'
      operationId: adminAibomOptions
      responses:
        '200':
          description: 참조 후보
          content:
            application/json:
              schema:
                type: object
                required: [datasets, references]
                properties:
                  datasets:
                    type: array
                    items:
                      type: object
                      properties:
                        identity: { type: string }
                        name: { type: string }
                        role: { type: array, items: { type: string, enum: [pretraining, finetuning, evaluation] } }
                  references:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: integer }
                        type: { type: string }
                        uri: { type: string }
                        title: { type: [string, 'null'] }
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /api/v1/admin/model-research/stream:
    post:
      tags: [admin]
      summary: '[관리자] AI 모델 리서치 (SSE 스트리밍)'
      description: |
        OpenRouter(웹 검색 도구 포함)를 호출해 모델 메타데이터 초안과 근거(evidence) 맵을 **Server-Sent Events** 로 스트리밍합니다.
        **DB에는 쓰지 않고 읽기만 합니다.** 프롬프트에 기존 카탈로그의 `familyKey`/`familyName`/`familyDeveloper` 목록, `relationshipType` 어휘, 모델 ID 목록을 주입해
        초안이 기존 식별자를 재사용하도록 합니다. DB를 읽지 못하면 컨텍스트 없이 조사하고 `warnings` 에 한국어 경고를 추가합니다.
        `complete` 이벤트의 `result.draft` 는 `AdminModelWrite` 와 호환되며, 검토 후 `POST /api/v1/admin/models` 에 그대로 사용할 수 있습니다.
        `bomFormat`/`bomSpecVersion`/`bomSerialNumber`/`bomVersion` 은 웹 조사 대상이 아니므로 서버가 항상 `null` 로 채웁니다.
        `result.status` 가 `ambiguous` 또는 `not_found` 이면 `warnings` 첫 항목에 그 사유가 들어가고, `not_found` 인 초안은 식별자 외 값이 모두 `null` 입니다.

        환경 변수: `OPENROUTER_API_KEY`(필수), `OPENROUTER_MODEL`, `OPENROUTER_BASE_URL`, `OPENROUTER_HTTP_REFERER`, `OPENROUTER_TIMEOUT_SECONDS`(기본 120초)

        HTTP 상태는 항상 `200` 이며, 오류(API 키 미설정, 타임아웃, 429/401/403, 응답 검증 실패 등)는 `error` 이벤트로 전달됩니다.
        응답 헤더: `Cache-Control: no-cache, no-transform`, `X-Accel-Buffering: no`

        이벤트 종류 (`event:` 필드):

        | event | data | 설명 |
        |---|---|---|
        | `status` | `{stage, message}` | `researching` → `validating` 단계 알림 (병합은 `merging` → `validating`) |
        | `provider_event` | OpenRouter 원본 chunk (reasoning 필드 제거) | 파싱 불가 시 `{type: "unparsed", data}` |
        | `content` | `{delta, characters}` | 모델 응답 텍스트 증분과 누적 글자 수 |
        | `activity` | `{kind: "citation"\|"tool", message, toolCalls?}` | 출처 확인 / 웹 검색 도구 실행 알림 |
        | `complete` | `{result: ModelResearchResult}` | 최종 검증된 결과 (마지막 이벤트) |
        | `error` | `{message}` | 오류 (이후 스트림 종료) |

        data 페이로드 스키마는 `ResearchStreamEvent` 를 참고하세요.
      operationId: adminStreamModelResearch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelResearchRequest'
      responses:
        '200':
          description: SSE 스트림
          headers:
            Cache-Control:
              schema: { type: string }
              example: no-cache, no-transform
            X-Accel-Buffering:
              schema: { type: string }
              example: 'no'
          content:
            text/event-stream:
              schema:
                type: string
                description: '`event: <name>` / `data: <json>` 프레임의 연속'
              example: |
                event: status
                data: {"stage": "researching", "message": "웹 자료를 검색하고 응답을 분석하고 있습니다."}

                event: activity
                data: {"kind": "tool", "message": "웹 검색 도구를 실행하고 있습니다.", "toolCalls": []}

                event: content
                data: {"delta": "{\"draft\":", "characters": 9}

                event: status
                data: {"stage": "validating", "message": "수집한 정보를 검증하고 초안을 정리하고 있습니다."}

                event: complete
                data: {"result": {"status": "found", "source": "huggingface", "draft": {"modelId": "Qwen/Qwen2.5-7B-Instruct"}, "parentCandidates": [], "fieldEvidence": [], "sources": [], "warnings": [], "research": {}}}
        '422':
          $ref: '#/components/responses/ValidationError'
  /api/v1/admin/model-research/results:
    get:
      tags: [admin]
      summary: '[관리자] 저장된 AI 조사 결과 조회'
      description: |
        요청 모델명(`modelName`, 공백 정규화 후 비교) 기준으로 `model_research_result` 테이블에 저장된
        huggingface / paper / merged 결과를 반환합니다. 프론트는 조사 시작 전에 이 API로 이전 결과 유무를 확인하고,
        있으면 "저장된 결과 불러오기 / 새로 조사하기" 를 선택하게 합니다.
      operationId: adminGetResearchResults
      parameters:
        - name: modelName
          in: query
          required: true
          schema: { type: string, minLength: 2, maxLength: 512 }
      responses:
        '200':
          description: '저장된 결과 (없으면 `exists: false`, 각 source 는 null)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredResearchResults'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
    post:
      tags: [admin]
      summary: '[관리자] AI 조사 결과 저장'
      description: |
        조사 결과를 (modelName, source) 단위로 upsert 합니다. `replacePrevious: true` 이면 같은 modelName 의
        기존 행을 모두 삭제한 뒤 저장합니다. 프론트는 두 조사가 끝난 시점에 `replacePrevious: true` 로 huggingface/paper 를
        저장하고, AI 병합이 끝나면 `replacePrevious: false` 로 merged 를 추가 저장합니다.
      operationId: adminSaveResearchResults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveResearchResultsRequest'
      responses:
        '200':
          description: 저장 후 현재 저장 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredResearchResults'
        '422':
          $ref: '#/components/responses/ValidationError'
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
    delete:
      tags: [admin]
      summary: '[관리자] 저장된 AI 조사 결과 삭제'
      operationId: adminDeleteResearchResults
      parameters:
        - name: modelName
          in: query
          required: true
          schema: { type: string }
      responses:
        '204':
          description: 삭제됨 (없어도 204)
        '503':
          $ref: '#/components/responses/DatabaseUnavailable'
  /api/v1/admin/model-research/merge/stream:
    post:
      tags: [admin]
      summary: '[관리자] AI 조사 결과 병합 (SSE 스트리밍)'
      description: |
        `/model-research/stream` 을 `source=huggingface`, `source=paper` 로 각각 호출해 얻은 두 `ModelResearchResult` 를
        OpenRouter(웹 검색 없음)에 넘겨 **하나의 초안으로 병합**합니다. 항목별로 근거가 강한 값을 고르고, 선택 이유(`fieldDecisions`)와
        충돌 경고를 한국어로 남깁니다. 저장소 사실(modelId, packageUrl, artifactFormat 등)은 Hugging Face, 패밀리 사실(familyDeveloper,
        familyReleaseDate 등)은 논문이 명시한 경우 논문을 우선하되, fieldEvidence 신뢰도가 더 높은 쪽이 있으면 그쪽을 따릅니다.
        `description` 만 두 결과를 결합(`combined`)할 수 있고 나머지 값은 한쪽 값을 그대로 사용합니다.

        DB는 읽기만 하며(카탈로그 컨텍스트 주입), 결과는 저장하지 않습니다. 두 결과 중 하나만 있어도 호출할 수 있지만,
        프론트는 한 경로만 성공하면 병합을 생략하고 해당 결과를 그대로 사용합니다.

        이벤트 종류는 `/model-research/stream` 과 같고, `status.stage` 는 `merging` → `validating` 입니다.
        `complete` 이벤트의 `result` 는 `ModelMergeResult` 입니다.
      operationId: adminStreamModelMerge
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MergeResearchRequest'
      responses:
        '200':
          description: SSE 스트림
          content:
            text/event-stream:
              schema:
                type: string
              example: |
                event: status
                data: {"stage": "merging", "message": "두 조사 결과를 항목별로 비교하고 병합하고 있습니다."}

                event: complete
                data: {"result": {"status": "found", "source": "merged", "draft": {"modelId": "Qwen/Qwen2.5-7B-Instruct"}, "fieldDecisions": [{"field": "familyDeveloper", "chosen": "paper", "rationale": "논문 저자 소속이 명시되어 있습니다."}], "parentCandidates": [], "sources": [], "warnings": [], "research": {}}}
        '422':
          $ref: '#/components/responses/ValidationError'
components:
  parameters:
    FamilyKey:
      name: familyKey
      in: path
      required: true
      description: 패밀리 식별자 (`model_info.family_key`)
      schema:
        type: string
        maxLength: 100
      example: qwen2.5-7b
    ModelId:
      name: modelId
      in: path
      required: true
      description: '`namespace/model-name` 형식의 모델 ID. 슬래시를 포함하며 경로 나머지 전체가 매칭됩니다.'
      schema:
        type: string
        maxLength: 512
      example: meta-llama/Llama-3.1-8B-Instruct
  responses:
    DatabaseUnavailable:
      description: 모델 카탈로그 DB 연결 불가
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorDetail'
          example: { detail: The model catalog database is unavailable. }
    ValidationError:
      description: 요청 본문 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
  schemas:
...
    ErrorDetail:
      type: object
      required: [detail]
      properties:
        detail:
          type: string
    HTTPValidationError:
      type: object
      description: FastAPI 표준 Pydantic 검증 오류
      properties:
        detail:
          type: array
          items:
            type: object
            required: [loc, msg, type]
            properties:
              loc:
                type: array
                items:
                  anyOf: [{ type: string }, { type: integer }]
              msg: { type: string }
              type: { type: string }
...
    FamilyIndex:
      type: object
      required: [items, summary]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/FamilyIndexItem'
        summary:
          $ref: '#/components/schemas/FamilyIndexSummary'
      example:
        items:
          - familyKey: qwen2.5-7b
            familyName: Qwen2.5-7B
            releaseDate: '2024-09-19'
            modelCount: 5
            rootCount: 2
            derivedCount: 3
            defaultNodeId: Qwen/Qwen2.5-7B-Instruct
        summary:
          familyCount: 1
          modelCount: 5
          rootCount: 2
          derivedCount: 3
          transformationTypeCount: 3
          latestRelease: '2024-09-19'
...
    FamilyHierarchy:
      type: object
      description: |
        패밀리 계층 트리. `details_json.family_config` 의 키가 그대로 병합되므로 추가 속성이 존재할 수 있습니다.
        이 스키마는 **snake_case** 입니다.
      required: [id, label, root_nodes, derived_models, default_node_id]
      properties:
        id:
          type: string
          description: familyKey 와 동일
        label:
          type: string
          description: familyName 과 동일
        default_node_id:
          type: [string, 'null']
        document_node:
          type: object
          description: family_config 가 없을 때 생성되는 기본 문서 노드
          properties:
            id: { type: string, example: 'document:qwen2.5-7b' }
            title: { type: string }
            subtitle: { type: string }
        root_family:
          type: object
          additionalProperties: true
          description: '`details_json.family` 원본 (fallback 시)'
        root_nodes:
          type: array
          items:
            $ref: '#/components/schemas/HierarchyRootNode'
        derived_models:
          type: array
          items:
            $ref: '#/components/schemas/HierarchyDerivedModel'
      additionalProperties: true
...
    AdminModel:
      type: object
      description: 관리자용 평탄화 모델 레코드. `AdminModelWrite` 의 모든 필드에 계산된 `hierarchyDepth` 가 추가됩니다.
      allOf:
        - $ref: '#/components/schemas/AdminModelWrite'
        - type: object
          required: [hierarchyDepth]
          properties:
            checklistScore:
              type: [number, 'null']
              description: 응답에서는 항상 number 로 직렬화
            checklistPenaltyFactor:
              type: [number, 'null']
            hierarchyDepth:
              type: integer
              minimum: 0
              description: 서버가 계산한 트리 깊이 (루트 = 0)

    # ---------- admin / research ----------
...
    MergeResearchRequest:
      type: object
      required: [modelName, results]
      properties:
        modelName:
          type: string
          minLength: 2
          maxLength: 512
        results:
          type: object
          description: 경로별 조사 결과. 최소 한 개는 null 이 아니어야 합니다.
          properties:
            huggingface:
              oneOf: [{ $ref: '#/components/schemas/ModelResearchResult' }, { type: 'null' }]
            paper:
              oneOf: [{ $ref: '#/components/schemas/ModelResearchResult' }, { type: 'null' }]
...